Campbell function

Computation of the leaf angle distribution function value (freq) Ellipsoidal distribution function caracterised by the average leaf inclination angle in degree (ala) Campbell 1986

Usage

campbell(ala)

Arguments

ala

average leaf angle

Value

foliar_distrib list. lidf and litab

Details

This function is used by the fourSAIL and fourSAIL2 model