Gaussion distribution

Gaussion distribution

Usage

gauss_byMin_Max(n, m, s, lwr, upr, nnorm)

Arguments

n

number of cases in numeric format

m

mean value in numeric format

s

standard deviation in numeric format

lwr

min value in numeric format

upr

max value in numeric format

nnorm

number of cases * a number (numeric format). This value is taken by the function to take values in the selected range (min-max) upt to have the complete the number of cases

Value

a gaussian dist