PROSPECT-PRO

PROSPECT-PRO

Usage

prospect_PRO(N, Cab, Car, Anth, Cbrown, EWT, LMA, alpha, Prot, CBC)

Arguments

N

numeric. Leaf structure parameter

Cab

numeric. Chlorophyll content (microg.cm-2)

Car

numeric. Carotenoid content (microg.cm-2)

Anth

numeric. Anthocyain content (microg.cm-2)

Cbrown

numeric. Brown pigment content (Arbitrary units)

EWT

numeric. Equivalent Water Thickness (g.cm-2)

LMA

numeric. Leaf Mass per Area (g.cm-2)

alpha

numeric. Solid angle for incident light at surface of leaf

Prot

numeric. protein content (g.cm-2)

CBC

numeric. NonProtCarbon-based constituent content (g.cm-2)

Value

List of lambda with leaf directional-hemisphrical reflectance and transmittance

References

Féret J-B, Gitelson AA, Noble SD & Jacquemoud S, 2017. PROSPECT-D: Towards modeling leaf optical properties through a complete lifecycle. Remote Sensing of Environment, 193, 204–215. https://doi.org/10.1016/j.rse.2017.03.004

Féret, J.B., Berger, K., de Boissieu, F., Malenovský, Z., 2021. PROSPECT-PRO for estimating content of nitrogen-containing leaf proteins and other carbon-based constituents. Remote Sens. Environ. 252. https://doi.org/10.1016/j.rse.2020.112173

Jacquemoud S, Baret F, Hanocq J-F, 1992. Modeling spectral and bidirectional soil reflectance. Remote Sensing of Environment, 41, 123–132. https://doi.org/10.1016/0034-4257(92)90072-R

Jacquemoud, S., Baret, F., 1990. PROSPECT: a model of leaf optical properties spectra. Remote Sens. Environ. 34, 75–91. https://doi.org/10.1016/0034-4257 (90)90100-Z.

Authors:

Jean-Baptiste Feret (jb.feret@teledetection.fr)

Frédéric Baret (baret@avignon.inra.fr)

Stephane Jacquemoud (jacquemoud@ipgp.fr)

This function includes numerical optimizations proposed in the FLUSPECT code

Authors:

Wout Verhoef

Christiaan van der Tol (c.vandertol@utwente.nl)

Joris Timmermans